Config 타입 types.ts로 이동, ActionHandling 로직 action.ts로 이동
