+++ frontend/public/components/hypercloud/chatbot/action.ts
@@ -1,13 +1,25 @@
 import { history } from '@console/internal/components/utils';
-import { Message, MessageContent } from './types';
+import { Config, Message, MessageName, PayloadType, QuickReplyPayload } from './types';
 
-export class Action {
+export class ActionFactory {
   static createActionHandler(message: Message) {
-    if (!message.payload || !message.payload.type || !message.payload.payload) {
-      return new BaseActionHandler();
+    switch (message?.name) {
+      case MessageName.WEBCHAT_LOADED:
+        return new WebchatLoadedActionHandler();
+      case MessageName.WEBCHAT_OPENED:
+        return new WebchatOpenedActionHandler();
+      case MessageName.WEBCHAT_CLOSED:
+        return new WebchatClosedActionHandler();
+      case MessageName.WEBCHAT_READY:
+        return new WebchatReadyActionHandler();
+      default:
+        switch (message?.payload?.type) {
+          case PayloadType.QUICK_REPLY:
+            return new RouteActionHandler(message.payload);
+          default:
+            return new BaseActionHandler();
+        }
     }
-    // 라우팅 액션 이외의 액션이 생길 경우 핸들러 추가
-    return new RouteActionHandler(message.payload);
   }
 }
 
@@ -21,11 +33,40 @@ class BaseActionHandler implements ActionHandler {
   }
 }
 
+// Triggered when the webchat is loaded and ready to be opened
+class WebchatLoadedActionHandler implements ActionHandler {
+  execute(): void {}
+}
+
+// Triggered when the webchat button bubble is clicked
+class WebchatOpenedActionHandler implements ActionHandler {
+  execute(): void {
+    const chatbotConfig: Config = {
+      botName: '하콘', // TODO: i18n 적용
+      locale: window.localStorage.getItem('i18nextLng'),
+      showPoweredBy: false,
+    };
+    window.botpressWebChat.configure(chatbotConfig);
+  }
+}
+
+// Triggered when the webchat close button is clicked
+class WebchatClosedActionHandler implements ActionHandler {
+  execute(): void {}
+}
+
+// Triggered when the webchat is ready to accept events, like proactive triggers
+class WebchatReadyActionHandler implements ActionHandler {
+  execute(): void {
+    showWelcomeMsg();
+  }
+}
+
 class RouteActionHandler implements ActionHandler {
-  private payload: MessageContent;
+  private payload: QuickReplyPayload;
   private path: string = '';
 
-  constructor(payload: MessageContent) {
+  constructor(payload: QuickReplyPayload) {
     this.payload = payload;
     this.setupRoute();
   }
@@ -45,3 +86,11 @@ class RouteActionHandler implements ActionHandler {
     this.path && history.push(this.path);
   }
 }
+
+const showWelcomeMsg = () => {
+  window.botpressWebChat.sendEvent({
+    type: 'proactive-trigger',
+    channel: 'web',
+    payload: { text: 'fake message' },
+  });
+};
